% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden_af.R
\name{burden_af}
\alias{burden_af}
\title{Attributable fraction}
\usage{
burden_af(RR = 1.062, unit = 10, pm_concentration = 1)
}
\arguments{
\item{RR}{A number. The relative risk from an epidemiological study}

\item{unit}{A number. The unit change in PM2.5 concentration related to RR}

\item{pm_concentration}{A number. The change in PM2.5 of interest.}
}
\value{
A number. The fraction of total mortality attributable to the risk factor
}
\description{
Attributable fraction
}
\examples{
burden_af(RR = 1.14, pm_concentration = 0.5)
}
