% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{burden_le}
\alias{burden_le}
\title{An implementation of the IOM life expectancy spreadsheets}
\usage{
burden_le(demog_data, min_age_at_risk = 30, pm_concentration = 1,
  RR = 1.06, unit = 10, neonatal_deaths = TRUE)
}
\arguments{
\item{demog_data}{A data frame with columns of headed "age" (the age at which each age group begins), "population" (the size of the population) and "deaths" (the number of deaths in the population).}

\item{min_age_at_risk}{Numeric vector. The lowest age susceptible to air pollution.}

\item{pm_concentration}{A number. The population weighted-mean PM2.5 concentration of interest.}

\item{RR}{A number. Specifies the relative risk from an epidemiologiccal study.}

\item{unit}{A number. Speficies the unit change associated with the relative risk (RR).}

\item{neonatal_deaths}{Logical. Are neonatal deaths included?}

\item{start_age}{A numeric vector. The starting age of each age group}
}
\value{
A list of 3 elements:
\itemize{
  \item{The difference (between the baseline and impacted population) in:}
    \itemize{
      \item{Age-specific life expectancy (days)}
      \item{Age-specific life-years lived per 100,000}
      \item{Age-specific number of  deaths per 100,000}
      }
  \item{The baseline life table}
  \item{The impacted life table}}
}
\description{
Calculate the change in life expectancy associated with a reduction in risk of death
}
\examples{

# Estimate the loss of life expectancy assoicated with 1mcg/m3
# of PM2.5 using an abridged set of population and mortality data:
population <- subset(abridged_data,
                     time == 2011 & sex == "Persons" & measure == "Population",
                     select = c(age, value))
population <- population$value
deaths <- subset(abridged_data,
                 time == 2011 & sex == "Persons" & measure == "Deaths",
                 select = c(age, value))
start_age <- as.numeric(gsub(" .+", "", deaths$age))
deaths <- deaths$value
demog_data <- data.frame(age = start_age, population, deaths)
x <- burden_le(demog_data)
x[[1]][1, 2] # Change in LE at birth (days)
}

