% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{burden_yll}
\alias{burden_yll}
\title{Calculate the number of YLL}
\usage{
burden_yll(attributable_number, life_expectancy)
}
\arguments{
\item{attributable_number}{A numeric vector of age-specific attributable numbers}

\item{life_expectancy}{A numeriv vector of age-specific life expectancies}
}
\value{
A numeric vector of age_specific YLL
}
\description{
Calculate the number of YLL
}
\examples{
data(abridged_data)
year <- 2011
s <- "Persons"
population <- subset(abridged_data,
                     time == year & sex == s & measure == "Population",
                     select = c(age, value))
deaths <- subset(abridged_data,
                 time == year & sex == s & measure == "Deaths",
                 select = c(age, value))
start_age <- as.numeric(gsub(" .+", "", deaths$age))
le <- burden_le(population$value, deaths$value, start_age)
impacted_le <- le[[3]][, "ex"]
an <- burden_an(population$value, deaths$value, start_age)
yll <- burden_yll(an, impacted_le)
sum(yll)
}

