% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{calculate_burden}
\alias{calculate_burden}
\title{An implementation of the IOM life expectancy spreadsheets}
\usage{
calculate_burden(start_age, population, deaths, ages_at_risk = 30:90,
  pm_concentration = 1, RR = 1.06, unit = 10, neonatal_deaths = TRUE)
}
\arguments{
\item{start_age}{Numeric vector. The starting age of each age group}

\item{population}{A numeric vector. The age-specific population size.}

\item{deaths}{A numeric vector. The age-specific number of deaths.}

\item{ages_at_risk}{Numeric vector. The age groups exposed to risk.}

\item{pm_concentration}{A number. The population weighted-mean PM2.5 concentration of interest.}

\item{RR}{A number. Specifies the relative risk from an epidemiologiccal study.}

\item{neonatal_deaths}{Logical. Are neonatal deaths included?}
}
\value{
A numeric vector of length 120 specifying the impact factor over the next 120 years.

A list of interesting outputs from the burden calculations
}
\description{
Calculate the change in life expectancy associated with a reduction in risk of death
}
\examples{

current_burden(0:90, population, age)

}

