% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_factor.R
\name{impact_factor}
\alias{impact_factor}
\title{Calculate the Impact Factor in future years}
\usage{
impact_factor(pm_conc, lag_structure = 1, RR = 1.062, unit = 10)
}
\arguments{
\item{pm_conc}{A numeric vector of PM2.5 concentrations in future years. The final value
will be assumed to remain constant until the end of the 120 year  period.}

\item{lag_structure}{A numeric vector specifying the structure of any cessation lag.}

\item{RR}{A number. Specifies the relative risk from an epidemiologiccal study.}

\item{unit}{A number. The unit change associated with the relative  risk}
}
\value{
A numeric vector of length 120 specifying the impact factor over the next 120 years
}
\description{
produces a vector of the impact factor over 120 years.
}
\examples{

# No cessation lag. PM concentration falls by 1 mcg/m3 after 1 year
pm <- c(15, 14)
impact_factor(pm)

#US EPA cessation-lag structure
us_epa_lag <- cumsum(c(0.3, rep(0.5, 4), rep(0.2/15, 15)))
impact_factor(pm, us_epa_lag)

# 5 year linear cessation-lag. PM concentration falls by 0.1mcg/m3 per year for 10 years.
pm <- rev(seq(15, 14, 0.1))
linear_lag <- seq(0.1, 1, 5)
impact_factor(pm, linear_lag)
}

