% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{life_expectancy}
\alias{life_expectancy}
\title{Calculate life expectancy}
\usage{
life_expectancy(start_age, hazard, neonatal_deaths = TRUE)
}
\arguments{
\item{start_age}{A numeric vector. The first age (in years) in each age group.}

\item{hazard}{A numeric vector. The age specific hazard.}

\item{neonatal_deaths}{Logical. Are neonatal deaths included?}
}
\value{
A data frame of: \itemize{
\item{Age groups}
\item{Age-specific hazards}
\item{Population surviving from a cohort of 100,000}
\item{Survival probability}
\item{Number of deaths per 100,000}
\item{Number of life-years lived per 100,0000}
\item{life expectancy}}
}
\description{
Calculate life expectancy
}
\examples{

Example using data from
ons_data <- read.csv("ons_data.csv")
start_age <- c(0, 1, seq(5, 85, 5))
hazard <- ons_data$deaths / ons_data$pop
life_expectancy(start_age, hazard)

}

