% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{present_value}
\alias{present_value}
\title{Present value of life years gained in the future}
\usage{
present_value(life_years, vly, discount_rate = 0.03)
}
\arguments{
\item{life_years}{A numeric vector of life years gained by year}

\item{vly}{The value of a statistical life year}

\item{discount_rate}{The discount rate to be applied}
}
\value{
A numeric vector of the present value of life years gained in future years.
}
\description{
Present value of life years gained in the future
}
\examples{

population <- subset(single_year_data,
                     time == 2011 & sex == "Persons" & measure == "Population")
population <- population[, c("age", "value")]
population$age <- as.numeric(gsub(" .+", "", population$age))
colnames(population)[2] <- "population"
deaths <- subset(single_year_data,
                 time == 2011 & sex == "Persons"& measure == "Deaths")
deaths <- deaths[, "value"]
demog_data <- data.frame(population, deaths = deaths)

no_lag <- impact(demog_data)
ly_extended <- colSums(no_lag$ly_extended)

sum(present_value(ly_extended, 250000))

}

