% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_year_data}
\alias{single_year_data}
\title{Single year population and mortality data for Australia}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6666 rows and 5 columns.}
\usage{
single_year_data
}
\description{
A dataset containing population and mortality data in single year age groups 0 - 100+.
}
\details{
@format A dataframe with 19 rows and 3 variables
 \describe{
   \item{Measure}{Population or deaths}
   \item{sex}{Sex}
   \item{age}{Age}
   \item{time}{Year}
   \item{value}{Mid-year population or annual number of deaths}
 }
 @source \url{http://stat.data.abs.gov.au}
}
\keyword{datasets}

