% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{burden_af}
\alias{burden_af}
\title{Attributbale fractionA number. The population weighted-mean PM2.5 concentration of interest.}
\usage{
burden_af(RR = 1.062, unit = 10, pm_concentration = 1)
}
\arguments{
\item{RR}{A number. Specifies the relative risk from an epidemiologiccal study.}

\item{unit}{A number. Speficies the unit change associated with the relative risk (RR).}

\item{pm_concentration}{A number. The population weighted-mean PM2.5 concentration of interest.}
}
\value{
A number. The percent of mortality attributable to the risk factor
}
\description{
Calculate the attributable fraction related to a PM$_{2.5}$ concentration
}
\examples{

burden_af(RR = 1.14, pm_concentration = 0.5)
}

