% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{burden_le}
\alias{burden_le}
\title{An implementation of the IOM life expectancy spreadsheets}
\usage{
burden_le(population, deaths, start_age = 0:90, min_age_at_risk = 30,
  pm_concentration = 1, RR = 1.06, unit = 10, neonatal_deaths = TRUE)
}
\arguments{
\item{population}{A numeric vector. The age-specific population size.}

\item{deaths}{A numeric vector. The age-specific number of deaths.}

\item{start_age}{A numeric vector. The starting age of each age group}

\item{min_age_at_risk}{Numeric vector. The lowest age susceptible to air pollution.}

\item{pm_concentration}{A number. The population weighted-mean PM2.5 concentration of interest.}

\item{RR}{A number. Specifies the relative risk from an epidemiologiccal study.}

\item{unit}{A number. Speficies the unit change associated with the relative risk (RR).}

\item{neonatal_deaths}{Logical. Are neonatal deaths included?}
}
\value{
A numeric vector of length 120 specifying the impact factor over the next 120 years.

A list of 3 elements:
\itemize{
  \item{The difference (between the baseline and impacted population) in:}
    \itemize{
      \item{Age-specific life expectancy (days)}
      \item{Age-specific life-years lived per 100,000}
      \item{Age-specific number of  deaths per 100,000}
      }
  \item{The baseline life table}
  \item{The impacted life table}}
}
\description{
Calculate the change in life expectancy associated with a reduction in risk of death
}
\examples{

# Estimate the loss of life expectancy assoicated with 1mcg/m3
# of PM2.5 using an abridged set of population and mortality data:
population <- subset(abridged_data,
                     time == 2011 & sex == "Persons" & measure == "Population",
                     select = c(age, value))
population <- population$value
deaths <- subset(abridged_data,
                 time == 2011 & sex == "Persons" & measure == "Deaths",
                 select = c(age, value))
start_age <- as.numeric(gsub(" .+", "", deaths$age))
deaths <- deaths$value
x <- burden_le(population, deaths, start_age)
x[[1]][1, 2] # Change in LE at birth (days)
}

