% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden_yll.R
\name{burden_yll}
\alias{burden_yll}
\title{Calculate the number of YLL}
\usage{
burden_yll(attributable_number, life_expectancy)
}
\arguments{
\item{attributable_number}{A numeric vector of age-specific attributable numbers}

\item{life_expectancy}{A numeriv vector of age-specific life expectancies}
}
\value{
A numeric vector of age_specific YLL
}
\description{
Calculate the number of YLL
}
\examples{
population <- subset(abridged_data,
                     time == 2011 & sex == "Persons" & measure == "Population",
                     select = c(age, value))
population <- population$value
deaths <- subset(abridged_data,
                 time == 2011 & sex == "Persons" & measure == "Deaths",
                 select = c(age, value))
start_age <- as.numeric(gsub(" .+", "", deaths$age))
deaths <- deaths$value
demog_data <- data.frame(age = start_age, population, deaths)
le <- burden_le(demog_data)
impacted_le <- le[["impacted"]][, "ex"]
an <- burden_an(demog_data)
yll <- burden_yll(an, impacted_le)
sum(yll)
}
