% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_function.R
\name{impact}
\alias{impact}
\title{Calculate the impact of a reduction in PM concentration}
\usage{
impact(population, deaths, pm_concentration = c(20, 19), lag_structure = 1,
  RR = 1.06, unit = 10, max_age = 105, base_year = 2013)
}
\arguments{
\item{population}{A vector of the population in single-year age groups}

\item{deaths}{A vector of the total number of deaths in the population}

\item{pm_concentration}{The population-weighted PM2.5 concentration in each future year}

\item{lag_structure}{The structure of any cessation lag}

\item{RR}{The relative risk (or hazard ratio) to use from the assessment. This is taken
from epidemiological studies}

\item{unit}{The unit change in PM2.5 concentration for the RR}

\item{max_age}{The maximum age to use for the assessment}

\item{base_year}{The base year for the assessment}
}
\value{
A list of four matrices (with dimensions of age and calendar year) of:
\itemize{
\item{The difference in number of deaths over 120 years (extended population)}
\item{The difference in number of deaths over 120 years (current cohort)}
\item{The difference in number of life years over 120 years (extended population)}
\item{The difference in number of life years over 120 years (current cohort)}}
}
\description{
Calculate the impact of a reduction in PM concentration
}
\examples{

Reduce PM from 15 to 14 mcg/m3
pm <- c(15, 14)
epa_lag <- cumsum(c(0.3, rep(0.125, 4), rep(0.2/15, 15)))

impact(pop, deaths)
}

