% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact}
\alias{impact}
\title{An implementation of the IOMLIFET impact spreadsheets}
\usage{
impact(demog_data, delta_pm = 1, lag_structure = 1, RR = 1.06,
  unit = 10, max_age = 105, base_year = 2013, neonatal_deaths = TRUE)
}
\arguments{
\item{demog_data}{A data frame with three numeric columns headed "age" (the age at which each age group begins),
"population" (the size of the population) and "deaths" (the number of deaths in the population).}

\item{delta_pm}{The reduction in population-weighted PM2.5 concentration in
each future year.}

\item{lag_structure}{The structure of any cessation lag}

\item{RR}{The relative risk (or hazard ratio) to use from the assessment.
This is taken from epidemiological studies}

\item{unit}{The unit change in PM2.5 concentration for the RR}

\item{max_age}{The maximum age to use for the assessment}

\item{base_year}{The base year for the assessment}

\item{neonatal_deaths}{Logical. Are neonatal deaths included?}
}
\value{
A list of four matrices (with dimensions of age and calendar year)
  of:
    \itemize{
      \item{The difference in number of deaths over 120 years (extended population)}
      \item{The difference in number of deaths over 120 years (current cohort)}
      \item{The difference in number of Life-years over 120 years (extended population)}
      \item{The difference in number of Life-years over 120 years (current cohort)}
      }
}
\description{
Implements a life table method for estimating the mortality
  benefits of reduction in exposure to risk
}
\examples{

# Reduce PM by 1mcg/m3. No cessation lag. RR = 1.06 per 10mcg/m3
head(single_year_data)
population <- subset(single_year_data,
                     time == 2011 & sex == "Persons" & measure == "Population")
population <- population[, c("age", "value")]
population$age <- as.numeric(gsub(" .+", "", population$age))
colnames(population)[2] <- "population"
deaths <- subset(single_year_data,
                 time == 2011 & sex == "Persons"& measure == "Deaths")
deaths <- deaths[, "value"]
demog_data <- data.frame(population, deaths = deaths)

no_lag <- impact(demog_data)

# Plot the effect in the current cohort and the extended population
par(mfrow = c(2, 1), cex = 0.5, cex.main = 1.9,
    cex.lab = 1.5, cex.axis = 1.4)
plot(no_lag$year, no_lag$deaths_current,
     xlab = "Year", ylab = "Number",
     main = "Deaths avoided", type = "b")
points(no_lag$year, no_lag$deaths, col = "red")
abline(h = 0)
legend(2100, 700, c("Current cohort", "Total"),
       col=c("black", "red"), pch = 16, cex = 1.5, lty=2)
plot(no_lag$year, no_lag$ly_current,
     xlab = "Year", ylab = "Number",
     main = "Life-years gained", , type = "b")
points(no_lag$year, no_lag$ly, col = "red")
abline(h = 0)

# US EPA lag
lag <- cumsum(c(0.3, rep(0.5/4, 4), rep(0.2/15, 15)))
epa_lag <- impact(population, deaths, lag_structure = lag)

# Comparison of no lag and US EPA lag (Extended cohort)
plot(no_lag$year, no_lag$deaths,
     xlab = "Year", ylab = "Number",
     main = "Deaths avoided", type = "b")
points(epa_lag$year, epa_lag$deaths, col = "red", type = "b")
abline(h = 0)
legend(2100, 700, c("No lag", "US EPA lag"),
       col=c("black", "red"), pch = 21, cex = 2, lty=2)
plot(no_lag$year, no_lag$ly,
     xlab = "Year", ylab = "Number",
     main = "Life-years gained", , type = "b")
points(epa_lag$year, epa_lag$ly, col = "red", type = "b")
abline(h = 0)

# Assuming PM takes 10 years to fall by 1mcg and US EPA cessation lag
pm <- seq(0.1, 1, 0.1)
slow_pm <- impact(population, deaths, delta_pm = pm, lag_structure = lag)

plot(no_lag$year, no_lag$deaths,
     xlab = "Year", ylab = "Number",
     main = "Deaths avoided", type = "b")
points(epa_lag$year, epa_lag$deaths, col = "red", type = "b")
points(slow_pm$year, slow_pm$deaths, col = "blue", type = "b")
abline(h = 0)
legend(2080, 700, c("No lag", "US EPA lag", "Lag and gradual fall in PM"),
       col=c("black", "red", "blue"), pch = 21, cex = 2, lty=2)
plot(no_lag$year, no_lag$ly,
     xlab = "Year", ylab = "Number",
     main = "Life-years gained", type = "b")
points(epa_lag$year, epa_lag$ly, col = "red", type = "b")
points(slow_pm$year, slow_pm$ly, col = "blue", type = "b")
#abline(h = 0)
}

