% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact_factor}
\alias{impact_factor}
\title{Calculate the Impact Factor in future years}
\usage{
impact_factor(delta_pm = 1, lag_structure = 1, RR = 1.06, unit = 10)
}
\arguments{
\item{delta_pm}{A numeric vector. The annual reduction from baseline of
future PM2.5 concentrations. The final value will be assumed to remain
constant until the end of the 120 year  period.}

\item{lag_structure}{A numeric vector specifying the structure of any
cessation lag.}

\item{RR}{A number. Specifies the relative risk from an epidemiologiccal
study.}

\item{unit}{A number. The unit change associated with the relative  risk}
}
\value{
A numeric vector of length 120 specifying the impact factor over the
  next 120 years
}
\description{
produces a vector of the impact factor over 120 years.
}
\examples{

# PM concentration reduced by 1 mcg/m3 in year 1, no cessation lag, RR = 1.06 per 10mcg/m3.
impact_factor()

#US EPA cessation-lag structure
epa_lag <- cumsum(c(0.3, rep(0.5, 4), rep(0.2/15, 15)))
impact_factor(lag_structure = epa_lag)

# 5 year linear cessation-lag. PM concentration falls by 0.1mcg/m3 per year for 10 years.
pm <- seq(0.1, 1, 0.1)
linear_lag <- seq(0.1, 1, 5)
impact_factor(pm, linear_lag)
}

